+++ cmd/accounts.js
@@ -1,38 +1,42 @@
 const help = require('./help');
 const {getLocalTime} = require('../util/wdate');
 const {getMoney} = require('../util/wsteem');
+const {to} = require('../util/wutil');
+const {question} = require('../util/wutil');
 
 const steem = require('steem');
 const dateFormat = require('dateformat');
 const ora = require('ora');
+const axios = require('axios');
 
 // 기본값
 const STEEM_AUTHOR = process.env.STEEM_AUTHOR;
 const STEEM_KEY_POSTING = process.env.STEEM_KEY_POSTING;
 
-let total_vesting_shares;
-let total_vesting_fund_steem;
-let spinner;
-
-function analysis(result){
-
-	// console.log(result)
+/*
+* 계정 정보를 분석한다
+* @param acc 계정정보 
+* @param total_vesting_shares total vesting shares 
+* @param total_vesting_fund_steem total fund steem 
+* @return 분석된 json 정보
+*/
+function analysis(acc, total_vesting_shares, total_vesting_fund_steem){
 
 	let profile = undefined;
 	try{
-		profile = JSON.parse(result.json_metadata);
+		profile = JSON.parse(acc.json_metadata);
 		profile=profile?profile.profile:undefined;
 	}catch(e){}
 	// let profile_image = profile?profile.profile_image:undefined;
-	let profile_image = `https://steemitimages.com/u/${result.name}/avatar`;
+	let profile_image = `https://steemitimages.com/u/${acc.name}/avatar`;
 	let profile_name = profile?profile.name:undefined;
 	let profile_about = profile?profile.about:undefined;
 	let profile_location = profile?profile.location:undefined;
 
-	let reputation = steem.formatter.reputation(result.reputation); // reputation
-  let vesting_shares = getMoney(result.vesting_shares); // vest
-  let received_vesting_shares = getMoney(result.received_vesting_shares); // vest
-  let delegated_vesting_shares = getMoney(result.delegated_vesting_shares); // vest
+	let reputation = steem.formatter.reputation(acc.reputation); // reputation
+  let vesting_shares = getMoney(acc.vesting_shares); // vest
+  let received_vesting_shares = getMoney(acc.received_vesting_shares); // vest
+  let delegated_vesting_shares = getMoney(acc.delegated_vesting_shares); // vest
 
   vesting_shares = vesting_shares + received_vesting_shares - delegated_vesting_shares;
   vesting_shares = steem.formatter.vestToSteem(vesting_shares, total_vesting_shares, total_vesting_fund_steem);
@@ -44,7 +48,7 @@ function analysis(result){
   delegated_vesting_shares = steem.formatter.vestToSteem(delegated_vesting_shares, total_vesting_shares, total_vesting_fund_steem);
   delegated_vesting_shares = Math.round(delegated_vesting_shares);  
 
-  let reward_sp = steem.formatter.vestToSteem(getMoney(result.reward_vesting_balance), total_vesting_shares, total_vesting_fund_steem);
+  let reward_sp = steem.formatter.vestToSteem(getMoney(acc.reward_vesting_balance), total_vesting_shares, total_vesting_fund_steem);
   reward_sp = reward_sp.toPrecision(4);
 
 	return [
@@ -52,13 +56,13 @@ function analysis(result){
 			en : 'profile',
 			kr : '프로필',
 			values : [
-				{val : result.id, en : 'id', kr : '아이디'},
-				{val : result.name, en : 'author', kr : '이름'},
+				{val : acc.id, en : 'id', kr : '아이디'},
+				{val : acc.name, en : 'author', kr : '이름'},
 				{val : profile_image, en : 'image', kr : '프로필사진'},
 				{val : profile_name, en : 'name', kr : '닉네임'},
 				{val : profile_about, en : 'about', kr : '정보'},
 				{val : profile_location, en : 'location', kr : '사는곳'},
-				{val : dateFormat(getLocalTime(result.created), 'yyyy-mm-dd HH:MM:ss'), en : 'location', kr : '계정생성일'},
+				{val : dateFormat(getLocalTime(acc.created), 'yyyy-mm-dd HH:MM:ss'), en : 'location', kr : '계정생성일'},
 				
 			]
 		},
@@ -70,58 +74,29 @@ function analysis(result){
 				{val : vesting_shares, en : 'vesting_shares', kr : '스파', subfix:'SP'},
 				{val : received_vesting_shares, en : 'received_vesting_shares', kr : '스파(임차)', subfix:'SP'},
 				{val : delegated_vesting_shares, en : 'delegated_vesting_shares', kr : '스파(임대)', subfix:'SP'},
-				{val : result.voting_power/100, en : 'voting_power', kr : '보팅파워', subfix:'%'},
-				{val : dateFormat(getLocalTime(result.last_vote_time), 'yyyy-mm-dd HH:MM:ss'), en : 'last_vote_time', kr : '최근투표일'},
+				{val : acc.voting_power/100, en : 'voting_power', kr : '보팅파워', subfix:'%'},
+				{val : dateFormat(getLocalTime(acc.last_vote_time), 'yyyy-mm-dd HH:MM:ss'), en : 'last_vote_time', kr : '최근투표일'},
 			]
 		},
 		{
 			en : 'money',
 			kr : '잔고',
 			values : [
-				{val : getMoney(result.balance), en : 'balance', kr : '스팀', subfix:'STEEM'},
-				{val : getMoney(result.sbd_balance), en : 'sbd_balance', kr : '스달', subfix:'SBD'}
+				{val : getMoney(acc.balance), en : 'balance', kr : '스팀', subfix:'STEEM'},
+				{val : getMoney(acc.sbd_balance), en : 'sbd_balance', kr : '스달', subfix:'SBD'}
 			]
 		},
 		{
 			en : 'reward',
 			kr : '보상',
 			values : [
-				{val : getMoney(result.reward_sbd_balance), en : 'reward_sbd_balance', kr : '스달', subfix:'SBD'},
-				{val : getMoney(result.reward_steem_balance), en : 'reward_steem_balance', kr : '스팀', subfix:'STEEM'},
-				{val : getMoney(result.reward_vesting_balance), en : 'reward_vesting_balance', kr : '베스트', subfix:'VEST'},
+				{val : getMoney(acc.reward_sbd_balance), en : 'reward_sbd_balance', kr : '스달', subfix:'SBD'},
+				{val : getMoney(acc.reward_steem_balance), en : 'reward_steem_balance', kr : '스팀', subfix:'STEEM'},
+				{val : getMoney(acc.reward_vesting_balance), en : 'reward_vesting_balance', kr : '베스트', subfix:'VEST'},
 				{val : reward_sp, en : 'reward_vesting_balance', kr : '스파', subfix:'SP'}
 			]
 		}
 	];	
-}
-
-/*
-* 보상을 요청한다
-* @param results 결과정보
-* @param author 저자 
-* @param posting 포스팅키 
-*/
-function claimRewards(results, author, posting){
-
-	// 요청 정보가 본인이며 기본 설정값이 존재하는지 확인 
-	if(results.length==1 && results[0].name==author && posting!=undefined){
-
-		let reward_steem_balance = results[0].reward_steem_balance;
-    let reward_sbd_balance = results[0].reward_sbd_balance;
-    let reward_vesting_balance = results[0].reward_vesting_balance;
-
-		let r1 = Number(reward_steem_balance.split(" ")[0]);
-    let r2 = Number(reward_sbd_balance.split(" ")[0]);
-    let r3 = Number(reward_vesting_balance.split(" ")[0]);
-
-    if (r1 == 0 && r2 == 0 && r3 == 0) {
-			return Promise.resolve(`${author} no claim`);
-    }else{
-    	spinner.start('claim reward');
-    	return steem.broadcast.claimRewardBalanceAsync(posting, author, reward_steem_balance, reward_sbd_balance, reward_vesting_balance);
-    }
-	}
-	return Promise.resolve('no claim');
 }
 
 /*
@@ -150,65 +125,163 @@ function initParams(args)
 	return args;
 }
 
+/*
+* 비동기 작업을 수행한다
+* @param account 계정명
+* @param wif 포스팅키
+*/ 
+let spinner;
+async function processAsyc(account, wif){
+
+	// 오류처리용
+  let err;
+
+  // 글로벌 설정 값 로드 -*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/
+  let global;
+  spinner = ora().start('loading global');
+  [err, global] = await to(steem.api.getDynamicGlobalPropertiesAsync());  
+  if(err){
+    // 오류처리
+    spinner.fail();
+    return Promise.reject(err.toString());
+  }
+  spinner.succeed(' ');
+  let total_vesting_shares = getMoney(global.total_vesting_shares);
+	let total_vesting_fund_steem = getMoney(global.total_vesting_fund_steem);
+
+	// 계정정보 확인 -*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/
+  let acc;
+  spinner = ora().start('loading account');
+  [err, acc] = await to(steem.api.getAccountsAsync([account]));  
+  if(acc){
+    if(acc.length!=1){
+      // 오류처리
+      spinner.fail();
+      return Promise.reject(`입력하신 ${account} 정보를 확인 바랍니다.`);
+    }
+  }else{
+    // 오류처리
+    spinner.fail();
+    return Promise.reject(err.toString());
+  }
+  acc = acc[0];
+  spinner.succeed(' ');
+
+  // 로드된 계정 값 분석 -*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/
+	let groups = analysis(acc, total_vesting_shares, total_vesting_fund_steem);
+	for(let g of groups){
+		console.log(`===== ${g.kr} (${g.en}) =====\n`);
+		for(let v of g.values){
+			console.log(`${v.kr}(${v.en}) : ${v.val?v.val:(v.subfix?0:'N/A')} ${v.subfix?v.subfix:''}`);
+		}
+		console.log(``);
+	}
+
+	// 스파 임차정보 확인 -*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/
+	if(getMoney(acc.received_vesting_shares)>0){
+		const AXIOS_CONFIG = {
+		  headers: {'user-agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/51.0.2704.103 Safari/537.36'}
+		};
+
+		// let account = 'ravenkim';
+		let DELEGATORS_API_URL = `https://happyukgo.com/api/steemit/delegators/?id=${account}&hash=3535ffa23344c25d2dcdc991fbdb60a5&_=`+new Date().getTime();
+		let delegators;
+		spinner = ora().start('loading delegator');
+		[err, delegators] = await to(axios.get(DELEGATORS_API_URL, AXIOS_CONFIG));
+		if(err){
+	    // 오류처리
+	    spinner.fail();
+	    return Promise.reject(err.toString());
+	  }else if(!delegators.data){
+	  	// 오류처리
+	    spinner.fail();
+	    return Promise.reject('임차 정보를 확인할 수 없습니다.');
+	  }
+	  spinner.succeed(' ');
+
+		// 임차 스파정보 출력
+	  console.log('===== 임차 (delegators) =====\n');
+	  for(let de of delegators.data){
+	  	let de_sp = de.sp.toFixed(0);
+	  	console.log(`임대받은 아이디 : ${de.delegator}, 날짜 : ${de.time}, 스파 : ${de_sp} SP`);
+	  }
+	  console.log(``);
+	}
+
+	// 스파 임대정보 확인 -*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/
+	if(getMoney(acc.delegated_vesting_shares)>0){
+		let delegatee;
+		spinner = ora().start('loading delegatee');
+	  [err, delegatee] = await to(steem.api.getVestingDelegationsAsync(account, null, 1000));  
+	  if(err){
+	    // 오류처리
+	    spinner.fail();
+	    return Promise.reject(err.toString());
+	  }
+	  spinner.succeed(' ');
+	  
+	  // 임대 스파정보 출력
+	  console.log('===== 임대 (delegatee) =====\n');
+	  for(let de of delegatee){
+	  	let de_date = dateFormat(getLocalTime(de.min_delegation_time),'yyyy-mm-dd HH:MM:ss');
+	  	let de_sp = steem.formatter.vestToSteem(getMoney(de.vesting_shares), total_vesting_shares, total_vesting_fund_steem).toFixed(0);
+	  	console.log(`임대해준 아이디 : ${de.delegatee}, 날짜 : ${de_date}, 스파 : ${de_sp} SP`);
+	  }
+	  console.log(``);
+	}
+	
+	// 보상청구 -*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/-*/
+	// 설정값 계정 일치 & 포스팅키 및 보상이 존재하는 경우 요청한다
+	if(STEEM_AUTHOR==account && wif){
+		let reward_steem_balance = acc.reward_steem_balance;
+    let reward_sbd_balance = acc.reward_sbd_balance;
+    let reward_vesting_balance = acc.reward_vesting_balance;
+
+		let r1 = getMoney(reward_steem_balance);
+    let r2 = getMoney(reward_sbd_balance);
+    let r3 = getMoney(reward_vesting_balance);
+
+    if(r1+r2+r3>0){
+    	let claim;
+		  spinner = ora().start('claim reward');
+		  [err, claim] = await to(steem.broadcast.claimRewardBalanceAsync(wif, account, reward_steem_balance, reward_sbd_balance, reward_vesting_balance));  
+		  if(err){
+		    // 오류처리
+		    spinner.fail();
+		    return Promise.reject(err.toString());
+		  }
+		  spinner.succeed(' ');
+    }
+	}
+
+	// 결과 확인
+  return Promise.resolve(`${account} 님의 계정정보 분석이 완료 되었습니다.`);
+}
+
 module.exports = (args)=>{
 
   // 파라미터 초기화
 	args = initParams(args);
 
 	// 입력 파라미터 유효성 검증 
-	if(args.length!=1 && args.length!=2){
+	if(args.length<1){
 		console.error('\n    [경고] 파라미터 오류  : 아래 메뉴얼을 참조 바랍니다');
 		help('accounts');
 		return;	
 	}
 
-	// (STEEM) 스피너 동작
-	spinner = ora().start('load properties');
+	let account = args[0];
+	let wif = args[1];
 
-	// 글로벌 설정 값 로드 
-	steem.api.getDynamicGlobalPropertiesAsync().then(result=>{
-			
-		spinner.succeed('load properties - success');
-
-    total_vesting_shares = getMoney(result.total_vesting_shares);
-    total_vesting_fund_steem = getMoney(result.total_vesting_fund_steem);
-
-    spinner.start('load accounts');
-    // (STEEM) 계정 목록 정보 로드
-    return steem.api.getAccountsAsync(args);
-    
-	}).catch(e=>{
-		// spinner.stop();
-		spinner.fail('load properties - fail');
-		console.error(`load properties - fail : `, e);
-	}).then(results=>{
-
-		spinner.succeed('load accounts - success\n');
-
-		// 로드된 계정 목록 값 분석 
-		for(let result of results){
-			let groups = analysis(result);
-			for(let g of groups){
-				console.log(`===== ${g.kr} (${g.en}) =====\n`);
-				for(let v of g.values){
-					console.log(`${v.kr}(${v.en}) : ${v.val?v.val:(v.subfix?0:'N/A')} ${v.subfix?v.subfix:''}`);
-				}
-				console.log(``);
-			}
-		}
-
-		// 보상을 요청한다
-		return claimRewards(results, STEEM_AUTHOR, STEEM_KEY_POSTING);
-
-	}).catch(e=>{
-		spinner.fail('load accounts - fail');
-		console.error(`load accounts - fail : `, e);
-	}).then(result=>{
-		if(spinner.isSpinning){
-			spinner.succeed('claim reward - success\n');	
-		}
-	}).catch(e=>{
-		spinner.fail('claim reward - fail');
-		console.error(`claim reward - fail : `, e);
+	processAsyc(account, wif)
+	.then(message=>{
+		console.log(message);
 	})
+	.catch(e=>{
+		console.log(`____________________________________________________________`);
+		console.error(e);
+		console.log(`____________________________________________________________`);
+		console.error('오류가 발생 했습니다.위쪽 라인을 참조 바랍니다.');
+		console.log(`____________________________________________________________`);
+	});
 };