+++ cmd/help.js
@@ -3,32 +3,40 @@ const menus = {
     steemkr [command] <options>
 
     (메인)
+    convert / ct  ............ 내부거래소 평균 시세로 SBD 로 STEEM 을 3.5 일 후 교환 / 취소 불가
     buysteem / bm  ........... 내부거래소에서 SBD 로 STEEM 을 구매
     buysbd / bd  ............. 내부거래소에서 STEEM 으로 SBD 를 구매
     buyscancel / bc  ......... 내부거래소에 등록된 목록 확인 및 취소
-    resteem / re  ........... 입력받은 주소 정보를 리스팀(리블로그) 한다. 7일이 지난 경우에도 가능
-    taglive / tl  ........... 입력받은 태그의 실시간 피드정보를 확인할 수 있다. 10초 단위 목록 정보 갱신
-    powerup / pw  ........... 내 계정의 스팀을 스팀파워로 전환한다(몽땅 한번에 전환함에 유의!)
-    block / bl  ............. 입력받은 계정을 차단한 계정목록을 볼 수 있다.
-    feed / fd ............... 입력받은 계정의 실시간 피드정보를 확인할 수 있다. 10초 단위 목록 정보 갱신
-    slb / sl ................ 입력받은 계정 및 기간 기준 스팀잇 라이프 벨런스를 보여준다.(글 작성 시간대)
-    accounts / ac ........... 입력받은 계정의 정보를 보여줌
-    price / pr .............. 오늘의 스팀 가격정보를 알려준다. (upbit 기준)
+    resteem / re  ............ 입력받은 주소 정보를 리스팀(리블로그) 한다. 7일이 지난 경우에도 가능
+    taglive / tl  ............ 입력받은 태그의 실시간 피드정보를 확인할 수 있다. 10초 단위 목록 정보 갱신
+    powerup / pw  ............ 내 계정의 스팀을 스팀파워로 전환한다(몽땅 한번에 전환함에 유의!)
+    block / bl  .............. 입력받은 계정을 차단한 계정목록을 볼 수 있다.
+    feed / fd ................ 입력받은 계정의 실시간 피드정보를 확인할 수 있다. 10초 단위 목록 정보 갱신
+    slb / sl ................. 입력받은 계정 및 기간 기준 스팀잇 라이프 벨런스를 보여준다.(글 작성 시간대)
+    accounts / ac ............ 입력받은 계정의 정보를 보여줌
+    price / pr ............... 오늘의 스팀 가격정보를 알려준다. (upbit 기준)
 
     (기타)
     version / vr ............ steemkr의 버전 정보
     help / hp ............... 도움말, help <options> 에서 options에 명령어를 입력하면 상세 도움말을 볼 수 있습니다
-
-    (비고)
-
-    <키값> .................... 설정값에서 해당 '키값' 기준으로 로딩합니다
-    [값] ...................... 반드시 필수로 입력 받아야 됩니다. 설정값에 설정 불가
-
+    
     $ steemkr help accounts ..... 계정관련 도움말을 볼 수 있습니다.
+    $ steemkr hp accounts ....... (줄임 명령어 사용) 계정관련 도움말을 볼 수 있습니다.
+  `,
+
+  convert: `
+    $ steemkr convert <STEEM_AUTHOR> <STEEM_KEY_ACTIVE>
+
+      * 내부거래소 평균 시세로 SBD 로 STEEM 을 3.5 일 후 교환 / 취소 불가
+
+    $ steemkr convert
+    $ steemkr convert 계정명 엑티브키
+    $ steemkr ct
+    $ steemkr ct 계정명 엑티브키
   `,
 
   buycancel: `
-    $ steemkr buysbd <STEEM_RESTEEM_AUTHOR> <STEEM_KEY_ACTIVE>
+    $ steemkr buysbd <STEEM_AUTHOR> <STEEM_KEY_ACTIVE>
 
       * 내부거래소에 등록한 목록을 확인 / 취소 할 수 있다.
 
@@ -39,7 +47,7 @@ const menus = {
   `,
 
   buysbd: `
-    $ steemkr buysbd <STEEM_RESTEEM_AUTHOR> <STEEM_KEY_ACTIVE> <STEEM_ORDER_LIMIT-option>
+    $ steemkr buysbd <STEEM_AUTHOR> <STEEM_KEY_ACTIVE> <STEEM_ORDER_LIMIT-option>
 
       * STEEM 로 SBD 구매를 진행한다
       * STEEM 가 0 이상 있는 경우에만 진행 가능함
@@ -51,7 +59,7 @@ const menus = {
   `,
 
   buysteem: `
-    $ steemkr buysteem <STEEM_RESTEEM_AUTHOR> <STEEM_KEY_ACTIVE> <STEEM_ORDER_LIMIT-option>
+    $ steemkr buysteem <STEEM_AUTHOR> <STEEM_KEY_ACTIVE> <STEEM_ORDER_LIMIT-option>
 
       * SBD 로 STEEM 구매를 진행한다
       * SBD 가 0 이상 있는 경우에만 진행 가능함